requirements.txt
# ── Core framework ────────────────────────────────────────────────────────
fastapi==0.115.12
uvicorn[standard]==0.34.0

# ── Settings & validation ──────────────────────────────────────────────────
pydantic==2.11.1
pydantic-settings==2.8.1

# ── File handling ──────────────────────────────────────────────────────────
python-multipart==0.0.20   # Required for FastAPI file upload support
aiofiles==24.1.0           # Async file I/O

# ── Testing (optional but recommended) ────────────────────────────────────
httpx==0.28.1              # Async test client for FastAPI
pytest==8.3.5
pytest-asyncio==0.25.3

# LLM Integration
openai>=1.14.0

# Database & Analytics
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.1

# ── Extraction (Phase 2) ──────────────────────────────────────────────────
pypdf==5.3.0
python-docx==1.1.2

# ── Analysis (Phase 3) ────────────────────────────────────────────────────
# Using OpenRouter Responses API Beta via httpx and requests
requests==2.32.3
